增加新的测试用例

+++ jt808/src/main/java/ai/sangmado/jt808/protocol/message/content/JT808_Message_Content_0x8100.java
@@ -3,8 +3,8 @@ package ai.sangmado.jt808.protocol.message.content;
 import ai.sangmado.jt808.protocol.ISpecificationContext;
 import ai.sangmado.jt808.protocol.enums.JT808DeviceRegistrationResult;
 import ai.sangmado.jt808.protocol.enums.JT808MessageId;
-import ai.sangmado.jt808.protocol.message.codec.IJT808MessageBufferReader;
-import ai.sangmado.jt808.protocol.message.codec.IJT808MessageBufferWriter;
+import ai.sangmado.jt808.protocol.message.encoding.IJT808MessageBufferReader;
+import ai.sangmado.jt808.protocol.message.encoding.IJT808MessageBufferWriter;
 import lombok.*;
 
 /**
@@ -52,8 +52,7 @@ public class JT808_Message_Content_0x8100 extends JT808MessageContent {
     public void deserialize(ISpecificationContext ctx, IJT808MessageBufferReader reader) {
         setAckSerialNumber(reader.readWord());
 
-        JT808DeviceRegistrationResult registrationResult =
-                JT808DeviceRegistrationResult.cast((int) reader.readByte());
+        JT808DeviceRegistrationResult registrationResult = JT808DeviceRegistrationResult.cast(reader.readByte());
         setRegistrationResult(registrationResult);
 
         // 终端注册成功后才有鉴权码字段
